Split implementation and prepare for HeapItem type family
We now have the following modules:

 - Data.Heap
 - Data.Heap.Packed
 - Data.Heap.Raw

Data.Heap.Raw is the only one implemented in this patch and basically contains a
minimal priority heap based on the last version. The HeapPolicy class has been removed.

The plan is to implement the HeapItem type family in Data.Heap.Packed. The module will
re-export basic functionality from Data.Heap.Raw and 'overwrite' some of the functions.

Data.Heap is scheduled to re-export Data.Heap.Packed.